tweak: rename variables and refactor css
